Creates a Home page component
